+++ .gitignore
@@ -1,7 +1,30 @@
-node_modules
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# large files
 *.wasm
-content_dist/*
 sandbox_dist/*
 buildmodel/*
 models/*
-temp/*
+old_code/*
