+++ build.sh
@@ -1,0 +1,2 @@
+cross-env INLINE_RUNTIME_CHUNK=false react-scripts build
+npx webpack --env config=sandbox
