--- content_src/index.ts
@@ -1,1 +1,0 @@
-// TODO: Interactions with webpage
