--- manifest.json
@@ -1,51 +1,0 @@
-{
-    "manifest_version": 3,
-    "name": "Clinical Notes AI",
-    "description": "Simplifying your clinical notes.",
-    "version": "0.1",
-
-    "icons": {
-        "16": "/images/icon-16.png",
-        "32": "/images/icon-32.png",
-        "48": "/images/icon-48.png",
-        "128": "/images/icon-128.png"
-    },
-
-    "background": {
-        "service_worker": "service_worker.js"
-    },
-
-    "content_scripts": [
-        {
-            "matches": ["<all_urls>"],
-            "js": ["/content_dist/bundle.js"]
-        }
-    ],
-
-    "sandbox": {
-        "pages": [ "sandbox.html" ]
-    },
-
-    "web_accessible_resources": [
-        {
-            "resources": [ 
-                "src/models/*.json",
-                "src/models/*.onnx",
-                "dist/*.wasm"
-            ],
-            "matches": [
-                "http://*/*",
-                "https://*/*"
-            ]
-        }
-    ],
-
-    "minimum_chrome_version": "109",
-
-    "permissions": [
-        "storage",
-        "contextMenus",
-        "tabs",
-        "offscreen"
-    ]
-}
