--- offscreen.html
@@ -1,3 +1,0 @@
-<iframe src="../sandbox.html" id="sandbox" style="display: none">
-</iframe>
-<script src="offscreen.js"></script>
