+++ package.json
@@ -1,38 +1,45 @@
 {
-  "name": "opennotes-extension",
-  "version": "1.0.0",
+  "name": "opennotes-app",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "copy-webpack": "^2.0.5",
-    "fs": "0.0.1-security",
-    "lodash": "^4.17.21",
-    "onnxruntime-node": "^1.13.1",
-    "onnxruntime-web": "^1.13.1",
-    "tokenizers": "^0.12.0",
-    "ts-loader": "^9.4.2",
-    "typescript": "^4.9.4",
-    "util": "^0.12.5",
-    "webpack-dev-server": "^4.11.1"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.14",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "cross-env": "^7.0.3",
+    "onnxruntime-web": "^1.14.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4"
   },
-  "description": "Chrome extension for simplifying patient clinical notes",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "webpack": "sh webpack.sh"
+    "start": "react-scripts start",
+    "build": "sh build.sh",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/NickMannhardt/open-notes-app.git"
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/NickMannhardt/open-notes-app/issues"
-  },
-  "homepage": "https://github.com/NickMannhardt/open-notes-app#readme",
-  "devDependencies": {
-    "@types/chrome": "0.0.210",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.1"
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
