+++ public/offscreen.html
@@ -1,0 +1,3 @@
+<iframe src="sandbox.html" id="sandbox" style="display: none">
+</iframe>
+<script src="offscreen.js"></script>
