+++ public/sandbox.html
@@ -1,0 +1,3 @@
+<!DOCTYPE html>
+    <script src="sandbox_dist/bundle.js"></script>
+</html>
