--- sandbox.html
@@ -1,3 +1,0 @@
-<!DOCTYPE html>
-    <script src="sandbox_dist/bundle.js"></script>
-</html>
