--- webpack.sh
@@ -1,7 +1,0 @@
-#! /bin/bash
-
-configs=("content" "sandbox")
-
-for c in ${configs[@]}; do
-    npx webpack --env config=$c
-done
